# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * UNY CBT Trainer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SubskillResult {
  name: string;
  attempts: number;
  correct: number;
}

export interface SectionResult {
  section: string;
  total: number;
  correct: number;
  incorrect: number;
  unanswered: number;
  accuracy: number;
  subskills: SubskillResult[];
}

export interface AnalyzeInput {
  totalQuestions: number;
  correct: number;
  incorrect: number;
  unanswered: number;
  score: number;
  accuracy: number;
  sections: SectionResult[];
  /** Total number of past exam sessions */
  historyCount: number;
}

export interface AnalysisResult {
  /** Short overall performance summary */
  summary: string;
  /** Areas where the student performs well */
  strengths: string[];
  /** Areas that need improvement */
  weaknesses: string[];
  /** Specific actionable study tips */
  recommendations: string[];
  /** Estimated UNY readiness level (e.g. Siap, Perlu Latihan Lebih, Belum Siap) */
  readinessLevel: string;
  /** A motivational message for the student */
  motivationalMessage: string;
}

export interface ExplainInput {
  section: string;
  subskill: string;
  question: string;
  /** @nullable */
  passage?: string | null;
  options: string[];
  correctAnswer: number;
  /** @nullable */
  userAnswer?: number | null;
  basicExplanation: string;
}

export interface ExplainResult {
  /** Step-by-step detailed explanation in Bahasa Indonesia */
  detailedExplanation: string;
  /** The underlying concept or skill being tested */
  concept: string;
  /** A tip for similar questions in the future */
  tipForNext: string;
}

export interface GenerateQuestionsInput {
  /** Section ID: tpa, lit_id, lit_en, or mtk */
  section: string;
  /** Specific subskill to generate questions for */
  subskill: string;
  /** Difficulty: mudah, sedang, or sulit */
  difficulty: string;
  /** Number of questions to generate (max 10) */
  count?: number;
}

export interface GeneratedQuestion {
  question: string;
  /** @nullable */
  passage?: string | null;
  options: string[];
  correctAnswer: number;
  explanation: string;
  estimatedTime: number;
}

export interface GenerateQuestionsResult {
  questions: GeneratedQuestion[];
  section: string;
  subskill: string;
  difficulty: string;
}

export interface Question {
  id: string;
  section: string;
  subskill: string;
  difficulty: string;
  sourceType: string;
  question: string;
  /** @nullable */
  passage?: string | null;
  options: string[];
  correctAnswer: number;
  explanation: string;
  estimatedTime: number;
}

export interface QuestionResult {
  question: Question;
  /** @nullable */
  userAnswer: number | null;
  isCorrect: boolean;
  doubtful: boolean;
  timeSpent: number;
}

export interface SectionStat {
  total: number;
  correct: number;
  incorrect: number;
  unanswered: number;
  accuracy: number;
  avgTime: number;
}

export type AttemptInputSections = {[key: string]: SectionStat};

export interface AttemptInput {
  /** Client-generated attempt id (e.g. "exam-<timestamp>") */
  id: string;
  /** mini | full | materi | review | tryout */
  mode: string;
  /**
     * Set when mode is "tryout" — links the attempt to a tryout set
     * @nullable
     */
  tryoutSetCode?: string | null;
  title: string;
  totalQuestions: number;
  correct: number;
  incorrect: number;
  unanswered: number;
  score: number;
  accuracy: number;
  timeUsedSec: number;
  durationSec: number;
  sections: AttemptInputSections;
  questions: QuestionResult[];
}

export type AttemptSections = {[key: string]: SectionStat};

export interface Attempt {
  id: string;
  mode: string;
  /** @nullable */
  tryoutSetCode?: string | null;
  title: string;
  totalQuestions: number;
  correct: number;
  incorrect: number;
  unanswered: number;
  score: number;
  accuracy: number;
  timeUsedSec: number;
  durationSec: number;
  sections: AttemptSections;
  questions: QuestionResult[];
  completedAt: string;
}

export interface ExamFormatSection {
  sectionId: string;
  questionCount: number;
  durationSeconds: number;
  order: number;
}

export interface ExamFormat {
  id: number;
  code: string;
  label: string;
  /** @nullable */
  description?: string | null;
  sections: ExamFormatSection[];
  totalQuestions: number;
  totalDurationSeconds: number;
  /** @nullable */
  sourceNotes?: string | null;
  status: string;
}

export interface TryoutSetProgress {
  /** belum | sedang | selesai */
  status: string;
  /** @nullable */
  bestScore: number | null;
  attemptsCount: number;
  /** @nullable */
  lastAttemptId: string | null;
  /** @nullable */
  lastCompletedAt: string | null;
}

export interface TryoutSet {
  id: number;
  code: string;
  label: string;
  /** @nullable */
  description?: string | null;
  examFormat: ExamFormat;
  progress: TryoutSetProgress;
}

export interface TryoutSetItem {
  questionId: string;
  sectionId: string;
  subskill: string;
  difficulty: string;
  orderIndex: number;
}

export interface WrongAnswer {
  questionId: string;
  timesMissed: number;
  lastMissedAt: string;
}

/**
 * Anonymous device id (UUID), generated and persisted client-side
 */
export type AnonIdHeaderParameter = string;