# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UNY CBT Trainer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attempts:
    get:
      operationId: listAttempts
      tags: [progress]
      summary: List this device's completed exam attempts (history)
      parameters:
        - $ref: "#/components/parameters/AnonIdHeader"
      responses:
        "200":
          description: List of attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attempt"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAttempt
      tags: [progress]
      summary: Save a completed exam attempt and update the wrong-answer bank
      parameters:
        - $ref: "#/components/parameters/AnonIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttemptInput"
      responses:
        "200":
          description: Saved attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attempts/{id}:
    get:
      operationId: getAttempt
      tags: [progress]
      summary: Get a single attempt by id
      parameters:
        - $ref: "#/components/parameters/AnonIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    AnonIdHeader:
      name: X-Anon-Id
      in: header
      required: true
      description: Anonymous device id (UUID), generated and persisted client-side
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Question:
      type: object
      properties:
        id:
          type: string
        section:
          type: string
        subskill:
          type: string
        difficulty:
          type: string
        sourceType:
          type: string
        question:
          type: string
        passage:
          type: ["string", "null"]
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: integer
        explanation:
          type: string
        estimatedTime:
          type: integer
      required:
        - id
        - section
        - subskill
        - difficulty
        - sourceType
        - question
        - options
        - correctAnswer
        - explanation
        - estimatedTime
    QuestionResult:
      type: object
      properties:
        question:
          $ref: "#/components/schemas/Question"
        userAnswer:
          type: ["integer", "null"]
        isCorrect:
          type: boolean
        doubtful:
          type: boolean
        timeSpent:
          type: number
      required:
        - question
        - userAnswer
        - isCorrect
        - doubtful
        - timeSpent
    SectionStat:
      type: object
      properties:
        total:
          type: integer
        correct:
          type: integer
        incorrect:
          type: integer
        unanswered:
          type: integer
        accuracy:
          type: number
        avgTime:
          type: number
      required:
        - total
        - correct
        - incorrect
        - unanswered
        - accuracy
        - avgTime
    AttemptInput:
      type: object
      properties:
        id:
          type: string
          description: Client-generated attempt id (e.g. "exam-<timestamp>")
        mode:
          type: string
          description: "mini | full | materi | review | tryout"
        tryoutSetCode:
          type: ["string", "null"]
          description: Set when mode is "tryout" — links the attempt to a tryout set
        title:
          type: string
        totalQuestions:
          type: integer
        correct:
          type: integer
        incorrect:
          type: integer
        unanswered:
          type: integer
        score:
          type: integer
        accuracy:
          type: integer
        timeUsedSec:
          type: integer
        durationSec:
          type: integer
        sections:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/SectionStat"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResult"
      required:
        - id
        - mode
        - title
        - totalQuestions
        - correct
        - incorrect
        - unanswered
        - score
        - accuracy
        - timeUsedSec
        - durationSec
        - sections
        - questions
    Attempt:
      type: object
      properties:
        id:
          type: string
        mode:
          type: string
        tryoutSetCode:
          type: ["string", "null"]
        title:
          type: string
        totalQuestions:
          type: integer
        correct:
          type: integer
        incorrect:
          type: integer
        unanswered:
          type: integer
        score:
          type: integer
        accuracy:
          type: integer
        timeUsedSec:
          type: integer
        durationSec:
          type: integer
        sections:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/SectionStat"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResult"
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - mode
        - title
        - totalQuestions
        - correct
        - incorrect
        - unanswered
        - score
        - accuracy
        - timeUsedSec
        - durationSec
        - sections
        - questions
        - completedAt
...