# lib/api-zod/src/generated/types/attempt.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * UNY CBT Trainer API
 * OpenAPI spec version: 0.1.0
 */
import type { AttemptSections } from './attemptSections';
import type { QuestionResult } from './questionResult';

export interface Attempt {
  id: string;
  mode: string;
  /** @nullable */
  tryoutSetCode?: string | null;
  title: string;
  totalQuestions: number;
  correct: number;
  incorrect: number;
  unanswered: number;
  score: number;
  accuracy: number;
  timeUsedSec: number;
  durationSec: number;
  sections: AttemptSections;
  questions: QuestionResult[];
  completedAt: Date;
}