# lib/api-zod/src/generated/types/attemptInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * UNY CBT Trainer API
 * OpenAPI spec version: 0.1.0
 */
import type { AttemptInputSections } from './attemptInputSections';
import type { QuestionResult } from './questionResult';

export interface AttemptInput {
  /** Client-generated attempt id (e.g. "exam-<timestamp>") */
  id: string;
  /** mini | full | materi | review | tryout */
  mode: string;
  /**
     * Set when mode is "tryout" — links the attempt to a tryout set
     * @nullable
     */
  tryoutSetCode?: string | null;
  title: string;
  totalQuestions: number;
  correct: number;
  incorrect: number;
  unanswered: number;
  score: number;
  accuracy: number;
  timeUsedSec: number;
  durationSec: number;
  sections: AttemptInputSections;
  questions: QuestionResult[];
}