# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * UNY CBT Trainer API
 * OpenAPI spec version: 0.1.0
 */

export interface Question {
  id: string;
  section: string;
  subskill: string;
  difficulty: string;
  sourceType: string;
  question: string;
  /** @nullable */
  passage?: string | null;
  options: string[];
  correctAnswer: number;
  explanation: string;
  estimatedTime: number;
}