# lib/api-zod/src/generated/types/questionResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * UNY CBT Trainer API
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from './question';

export interface QuestionResult {
  question: Question;
  /** @nullable */
  userAnswer: number | null;
  isCorrect: boolean;
  doubtful: boolean;
  timeSpent: number;
}