# lib/api-zod/src/generated/types/sectionStat.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * UNY CBT Trainer API
 * OpenAPI spec version: 0.1.0
 */

export interface SectionStat {
  total: number;
  correct: number;
  incorrect: number;
  unanswered: number;
  accuracy: number;
  avgTime: number;
}